requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.8
iniconfig==2.0.0
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pdfrw==0.4
pillow==10.4.0
pluggy==1.5.0
praw==7.7.1
prawcore==2.4.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
